+++ sv-parser-pp/testcases/test4.sv
@@ -1,0 +1,16 @@
+`define disp(clk, exp, msg) \
+    always @(posedge clk) begin \
+        if (!(exp)) begin \
+            $display msg; \
+        end \
+    end \
+
+module a ();
+
+`disp(
+    clk,
+    !(a[i].b && c[i]),
+    ("xxx(()[]]{}}}", a[i].b, c[i])
+);
+
+endmodule